requirements.txt
Django==3.0.6
django-cors-headers==3.3.0
djangorestframework==3.10.3
PyMySQL==0.9.3
mysqlclient==2.1.1
unidecode==1.3.2
